// game/Classic/InformationDilemma/src/interface.d.ts
export interface ICreateParams {

    /** CreateParams round */
    round?: (number|null);

    /** CreateParams groupSize */
    groupSize?: (number|null);

    /** CreateParams rightReward */
    rightReward?: (number|null);

    /** CreateParams falseReward */
    falseReward?: (number|null);

    /** CreateParams nextPhaseKey */
    nextPhaseKey?: (string|null);
}